+++ src/app/_layout.tsx
@@ -40,7 +40,15 @@ function RootNavigator() {
         <Stack.Screen name="index" options={{ headerShown: false }} />
       </Stack.Protected>
       <Stack.Protected guard={isAuthenticated}>
-        <Stack.Screen name="posts" options={{ title: "Ieraksti" }} />
+        <Stack.Screen
+          name="posts"
+          options={{
+            title: "Ieraksti",
+            headerStyle: { backgroundColor: colors.white },
+            headerTintColor: colors.black,
+            headerShadowVisible: false,
+          }}
+        />
       </Stack.Protected>
     </Stack>
   );