+++ src/theme.ts
@@ -3,4 +3,5 @@ export const colors = {
   purple: "#663399",
   yellow: "#FFDE59",
   black: "#000000",
+  lightGray: "#E5E5E5",
 } as const;